% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stability.R
\name{C}
\alias{C}
\title{This function gives the stability constant of a multiresource job queue with service times exponentially distributed and homogeneous across customer classes}
\usage{
C(P)
}
\arguments{
\item{P}{array where index is the multidimensional class of the customer, indexed from 0, should have P[1,...,1]=0}
}
\value{
constant C in the stability condition lambda C<mu, where lambda is the input rate, and mu is the service rate
}
\description{
This function gives the stability constant of a multiresource job queue with service times exponentially distributed and homogeneous across customer classes
}
