\name{X}
\alias{X}
\docType{data}
\title{
Dataset with raw workload data from HPDC KRC RAS
}
\description{
Source data for workload of HPC of HPDC KRC RAS. More usable dataset
is HPC_KRC. This are raw times in sec. since 1 January 1970, for
tasks arrival times, start of execution times and end times.
}
\usage{data(X)}
\format{
  The format is:
 num [1:8499, 1:3] 1.24e+09 1.24e+09 1.24e+09 1.24e+09 1.24e+09 ...
}
\source{
http://cluster.krc.karelia.ru
}
\references{
http://cluster.krc.karelia.ru
}
\examples{
data(X)
}
\keyword{datasets}
