% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMC.R
\name{DMC}
\alias{DMC}
\title{Distributional Measure of Correlation}
\usage{
DMC(X, Y)
}
\arguments{
\item{X}{First variable (vector)}

\item{Y}{Second variable (vector)}
}
\value{
One value between -1 and 1, characterizing the dependence between the variables
}
\description{
This is a suggested by Dror Feitelson measure of correlation for dependent
variables, that may be successfully used to examine the datasets from a
High Performance Cluster logs
}
\examples{
data(HPC_KRC)
DMC(HPC_KRC$service[1:1000], HPC_KRC$cores_requested[1:1000])
}
\references{
http://interstat.statjournals.net/YEAR/2004/abstracts/0412001.php?Name=412001
}
